feat[BetResultGenerator]: 베팅 결과 계산하는 클래스 생성 및 Referee 수정
